streaming/pyproject.toml
[project]
name = "streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.13.0.92",
    "cobs>=1.2",
    "crcmod>=1.7",
    "protobuf>=5.0",
    "pyserial>=3.5",
]
